reduce complexity and loc
