+++ playwright/.auth/state.json
@@ -1,109 +1,4 @@
 {
-  "cookies": [
-    {
-      "name": "oauth_state",
-      "value": "0235d2e3-8721-4972-802f-30c005070390",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1771380938.913694,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "_cfuvid",
-      "value": "B965UELo7gHeD9DCNafjAHswyAoYOPtGvJgv_3WSukw-1771380339263-0.0.1.1-604800000",
-      "domain": ".workos.com",
-      "path": "/",
-      "expires": -1,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "None"
-    },
-    {
-      "name": "__wuid",
-      "value": "Fe26.2*1*4590387feae5bfac25bbe85cfd18bdad8e4217fa7475c5b48ef37daebe96f2b6*0swgjH3CyaZT09XzmJA9XQ*CqvjX3gbCCMWzTGmzvGAmVHFn_roMwt6604Ui8x0CYGs4Y46SUO2wQ0XcRLpNMuju47dOj0vRIHrXx5SvdVBfq1yOJuP0KqlpGDXEA5PF7c*1776564341706*aa364af5141863d78cf2f62004b3876f203132a2c180271d274d0ad8bdcd83e6*MmaSluHnuOwgAzApZAT-yp4qK0LjVhXrdCUXwMge828~2",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": 1776564341,
-      "httpOnly": false,
-      "secure": true,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "_dd_s",
-      "value": "aid=44a49dda-48d0-40a6-9452-313f0adc2446&rum=2&id=9c705376-c6db-4434-b08f-f9a35a644bd6&created=1771380340666&expire=1771381244870",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": 1802916344,
-      "httpOnly": false,
-      "secure": false,
-      "sameSite": "Strict"
-    },
-    {
-      "name": "session",
-      "value": "UapqihIkWVzMhx0kFGuhKJqvt",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": -1,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "access-token",
-      "value": "eyJhbGciOiJSUzI1NiIsImtpZCI6InNzb19vaWRjX2tleV9wYWlyXzAxSzJINURFTUZUVlBEWlNLSFpWQ0RHWkM4In0.eyJhdWQiOiJodHRwczovL2FwaS53b3Jrb3MuY29tL3gvYXV0aGtpdCIsImlzcyI6Imh0dHBzOi8vYXBpLndvcmtvcy5jb20veC9hdXRoa2l0Iiwic3ViIjoidXNlcl8wMUs5Q0ZROTNZQ0EwRDhBUDg1QVNEUVdLUiIsInNpZCI6InNlc3Npb25fMDFLSFE3WTJSM1dBRkJLQ0o4RjM5MVE3R0MiLCJqdGkiOiIwMUtIUTdZMlNIRllCNlMwSlBaQzBHM0NRNCIsImV4cCI6MTc3MTM4MDY0NSwiaWF0IjoxNzcxMzgwMzQ1fQ.cRi1Zb_ZsntuZxBOy8FmWI8TC_5gnuBVjwiVoDquHI3QO24TINDUF6aIL0Gqcm49FlHVcLD_y9j2OfpQ7j3eKzQ0kkvH21qjTYbGhsO0UFm6K3ZjhN79LGcF6APMWOC9DIA5qWfEEaH8NIEndnhJuj6G5OsaJgGz0iDUrkxwam07RwxciXQ-p_JeJznJBOBi5zxVjBuqHaNwy4VaIhXUCI-gI-zzmQosSBfcNRj7resT407K_o6uIQsfx92qxiNM_V0unAin8PiIeuOStFIZCzWOqeNN3lrmRlGn9RmRE4Nie5HhRrmFeDJB7gK1am5-N6WF_igS8w2rNOs-_FQ1eg",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": 1771380645.833231,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "__kduid",
-      "value": "Fe26.2*1*3c0cc83af28e2e7f7694418658f4405e6195cd52a7aa9bb787edae948130a6cf*OkfZPrwvHqESMCx1Rdp_yQ*FJwq-U01_5AACw29n-XdU0USVQymLrELnXSMx10yNBfOgPyC-MNDkeq4-aZmu2-OqxN7VDnbfCu2AvGAXJ2etkRMNQRU5A7_mKwWoUeXLZ5a-pagcbxuIQScpusEi3Gr*1802916345709*e3399f027bebc46f570f17283d07e0ad0d7a7b5a21b3cce7264fafa26b38cdda*FiqeKbNhXfBf_MoULOuYWoLj2Dp9wmy9Lt5qu3W4drs~2",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": 1802916345.833259,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "__recent_auth",
-      "value": "Fe26.2*1*2039b2cc8b68d427ca031cbe925cc826310980a2409dbb3b58bfd50e2f70f6f3*Ar5ulsSJHOz687Rj55SmBg*Ed_ITPGwwe2nqK09sR1Ok2vPhSceAEVbsNA25NQSwBMCRhLe1iTHDBqO1AqaYNeb02Ygx_qAFF7hgE5my9na3w*1771985145710*e79175339dacd299219a28d8c3926dc48554fa6cd0d47e448244f05ba31472b3*7GjJeb2MR1zQ6wnDUYiddoz_tAvvYU8DRJo66eB51j0~2",
-      "domain": "fastidious-fuchsia-08-staging.authkit.app",
-      "path": "/",
-      "expires": 1771985145.833287,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "Lax"
-    },
-    {
-      "name": "session",
-      "value": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyXzAxSzlDRlE5M1lDQTBEOEFQODVBU0RRV0tSIiwiZW1haWwiOiJzdGV2aW8ud29uZGVyQGdtYWlsLmNvbSIsIndvcmtvc0lkIjoic2Vzc2lvbl8wMUtIUTdZMlIzV0FGQktDSjhGMzkxUTdHQyIsImlhdCI6MTc3MTM4MDM0NiwiZXhwIjoxNzcxNzI1OTQ2LCJpc3MiOiJmaXQtd29ya291dC1hcHAiLCJhdWQiOiJmaXQtd29ya291dC1hcHAifQ.UYDC0sRZkHzw2fjfrvF_Pn8GAN5qEmIBJ_o7ACJhv18",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1771725946.216421,
-      "httpOnly": true,
-      "secure": true,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": [
-    {
-      "origin": "http://localhost:8787",
-      "localStorage": [
-        {
-          "name": "theme",
-          "value": "system"
-        },
-        {
-          "name": "auth_user",
-          "value": "{\"id\":\"user_01K9CFQ93YCA0D8AP85ASDQWKR\",\"email\":\"stevio.wonder@gmail.com\",\"name\":\"stevio.wonder\",\"cachedAt\":1771380346615}"
-        }
-      ]
-    }
-  ]
+  "cookies": [],
+  "origins": []
 }
