+++ src/lib/programs/config/index.ts
@@ -1,0 +1,9 @@
+export * from './megsquats';
+export * from './stronglifts';
+export * from './wendler531';
+export * from './madcow';
+export * from './candito';
+export * from './nsuns';
+export * from './sheiko';
+export * from './nuckols';
+export * from './jen-sinkler';
