+++ src/lib/programs/megsquats.ts
@@ -1,55 +1,10 @@
 import { roundToPlate } from './utils';
 import { generateWorkoutAccessories } from './accessory-data';
-import type { LiftType, OneRMValues, ProgramConfig, ProgramWorkout, ProgramAccessory } from './types';
-
-const megsquatsInfo = {
-  slug: 'stronger-by-the-day',
-  name: 'Stronger by the Day (Megsquats)',
-  description: 'A 12-week upper/lower split program designed specifically for women, featuring training max progression and glute-focused accessories.',
-  difficulty: 'beginner' as const,
-  daysPerWeek: 3,
-  estimatedWeeks: 12,
-  totalSessions: 36,
-  mainLifts: ['squat', 'bench', 'deadlift', 'ohp'] as LiftType[],
-  category: "women's" as const,
-};
-
-const megsquatsAccessories: ProgramAccessory[] = [
-  { accessoryId: 'hip-thrust', sets: 3, reps: 10, isRequired: true },
-  { accessoryId: 'romanian-dl', sets: 3, reps: 10, isRequired: false },
-  { accessoryId: 'leg-press', sets: 3, reps: 12, isRequired: false },
-  { accessoryId: 'leg-curls', sets: 3, reps: 12, isRequired: false },
-  { accessoryId: 'planks', sets: 3, reps: '30 sec', isRequired: false },
-];
-
-const upperAccessories: ProgramAccessory[] = [
-  { accessoryId: 'pullups', sets: 3, reps: 8, isRequired: false },
-  { accessoryId: 'lat-pulldowns', sets: 3, reps: 10, isRequired: false },
-  { accessoryId: 'tricep-pushdowns', sets: 3, reps: 12, isRequired: false },
-  { accessoryId: 'face-pulls', sets: 3, reps: 15, isRequired: false },
-  { accessoryId: 'lateral-raises', sets: 3, reps: 15, isRequired: false },
-];
-
-function getMegsquatsAccessories(_week: number, session: number): ProgramAccessory[] {
-  const isLowerA = session === 1;
-  const isUpperA = session === 2;
-
-  if (isLowerA) {
-    return megsquatsAccessories.filter(acc =>
-      ['leg-curls', 'planks'].includes(acc.accessoryId)
-    );
-  }
-
-  if (isUpperA) {
-    return upperAccessories.filter(acc =>
-      ['lat-pulldowns', 'lateral-raises'].includes(acc.accessoryId)
-    );
-  }
-
-  return megsquatsAccessories.filter(acc =>
-    ['hip-thrust', 'planks'].includes(acc.accessoryId)
-  );
-}
+import {
+  megsquatsInfo,
+  getMegsquatsAccessories,
+} from './config/megsquats';
+import type { LiftType, OneRMValues, ProgramConfig, ProgramWorkout } from './types';
 
 function getWavePercentage(week: number, setNumber: number): number {
   const waveData: Record<number, number[]> = {
